Add a new parameter bAllDirect to FetchStatusForSingleTarget and ReadRemainingItemsStatus.
This parameter determines if all the files/folders returned when fetching the status should have the direct flag set.
Since we just fetch the status of the common parent and set a filter if multiple files are selected, this param is needed or else the directly selected files won't even show up in the control!