The SVNStatusListCtrl now shows the "Ignore" context menu for items in the status "deleted". Previously it was only shown for "unversioned" items.
Since "deleted" items are scheduled for deletion it can make sense to put them on the ignore list to prevent them being added again in the future. Although svn:ignore gets updated the items are not removed from the list of items, because the "deleted" status is more important than the fact that they are ignored.